Sends only one notification per batch review

+++ ee/app/services/draft_notes/publish_service.rb
@@ -8,6 +8,11 @@ module DraftNotes
       else
         publish_draft_notes
       end
+
+      success
+    rescue ActiveRecord::RecordInvalid => e
+      message = "Unable to save #{e.record.class.name}: #{e.record.errors.full_messages.join(", ")} "
+      error(message)
     end
 
     private
@@ -20,7 +25,21 @@ module DraftNotes
     end
 
     def publish_draft_notes
-      draft_notes.each(&method(:create_note_from_draft))
+      return if draft_notes.empty?
+
+      if Feature.enabled?(:batch_review_notification, project)
+        review = Review.create!(author: current_user, merge_request: merge_request, project: project)
+
+        draft_notes.map do |draft_note|
+          draft_note.review = review
+          create_note_from_draft(draft_note)
+        end
+
+        notification_service.async.new_review(review)
+      else
+        draft_notes.each(&method(:create_note_from_draft))
+      end
+
       draft_notes.delete_all
 
       MergeRequests::ResolvedDiscussionNotificationService.new(project, current_user).execute(merge_request)
@@ -33,6 +52,8 @@ module DraftNotes
 
       note = Notes::CreateService.new(draft.project, draft.author, draft.publish_params).execute
       set_discussion_resolve_status(note, draft)
+
+      note
     end
 
     def set_discussion_resolve_status(note, draft_note)
